topic_list: Use UnreadCountBadge, replacing ad hoc private widget
The private widget, following the Figma, made the badge a bit more
compact, with 1px less padding on each side, and a font size smaller
by 1 (but with the same line height).

That was helpful for keeping the rows at uniform height regardless
of unreads (without text scaling anyway)...but I've accomplished
that by adding a minimum row height (40), in the previous commit.